+++ src/main/java/sample/data/jpa/web/ReviewController.java
@@ -1,12 +1,18 @@
 package sample.data.jpa.web;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import sample.data.jpa.domain.Review;
 import sample.data.jpa.service.ReviewDao;
+
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/review")
@@ -17,40 +23,40 @@ public class ReviewController {
 
     @RequestMapping("/create")
     @ResponseBody
-    public String create(Review review) {
-        String reviewId = "";
-        try {
+    public ResponseEntity<?> create(@RequestBody Review review) {
+        if (Objects.nonNull(review)) {
             reviewDao.save(review);
-            reviewId = String.valueOf(review.getId());
-        } catch (Exception ex) {
-            return "Error creating the review: " + ex.toString();
+            var headers = new HttpHeaders();
+            headers.add(HttpHeaders.ACCESS_CONTROL_EXPOSE_HEADERS, "X-Custom-ID");
+            headers.add("X-Custom-ID", String.valueOf(review.getId()));
+            headers.setContentType(MediaType.APPLICATION_JSON);
+            return ResponseEntity.ok().headers(headers).body(review);
+        } else {
+            return ResponseEntity.badRequest().body("Cannot add review to the DB - operation failed");
         }
-
-        return "Review successfully created with id = " + reviewId;
     }
 
     @RequestMapping("{id}/delete")
     @ResponseBody
-    public String delete(Long id) {
-        try {
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        var review = reviewDao.findById(id);
+        if (review.isPresent()) {
             reviewDao.deleteReviewById(id);
-        } catch (Exception ex) {
-            return "Error deleting the review: " + ex;
+            return ResponseEntity.ok("Review successfully deleted!");
+        } else {
+            return ResponseEntity.notFound().build();
         }
-        return "Review successfully deleted!";
     }
 
     @RequestMapping("{id}")
     @ResponseBody
-    public String getById(@PathVariable("id") Long id) {
-        String reviewId = "";
-        try {
-            Review review = reviewDao.findReviewById(id);
-            reviewId = String.valueOf(review.getId());
-        } catch (Exception ex) {
-            return "City not found";
+    public ResponseEntity<?> getById(@PathVariable("id") Long id) {
+        var review = reviewDao.findById(id);
+        if (review.isPresent()) {
+            return ResponseEntity.ok(review.get());
+        } else  {
+            return ResponseEntity.notFound().build();
         }
-        return "The city id is: " + reviewId;
     }
 }
